[NEW] Agregando la creacion de ordenes

+++ src/main/java/com/restaurant/system/backend_restaurant_system/controller/CategoryController.java
@@ -63,12 +63,12 @@ public class CategoryController {
         try {
             categoryService.updateCategory(id, categoryDTO);
             MessageDTO response = new MessageDTO();
-            response.setStatus((long) HttpStatus.OK.value());
+            response.setStatus(HttpStatus.OK.value());
             response.setMessage("Categoría actualizada exitosamente");
             return new ResponseEntity<>(response, HttpStatus.OK);
         } catch (RuntimeException e) {
             MessageDTO response = new MessageDTO();
-            response.setStatus((long) HttpStatus.NOT_FOUND.value());
+            response.setStatus(HttpStatus.NOT_FOUND.value());
             response.setMessage(e.getMessage());
             return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
         }
