+++ src/main/java/com/restaurant/system/backend_restaurant_system/controller/DishController.java
@@ -65,12 +65,12 @@ public class DishController {
         try {
             dishService.updateDish(id, DishDTO);
             MessageDTO response = new MessageDTO();
-            response.setStatus((long) HttpStatus.OK.value());
+            response.setStatus(HttpStatus.OK.value());
             response.setMessage("Plato actualizado exitosamente");
             return new ResponseEntity<>(response, HttpStatus.OK);
         } catch (RuntimeException e) {
             MessageDTO response = new MessageDTO();
-            response.setStatus((long) HttpStatus.NOT_FOUND.value());
+            response.setStatus(HttpStatus.NOT_FOUND.value());
             response.setMessage(e.getMessage());
             return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
         }
