+++ src/main/java/com/restaurant/system/backend_restaurant_system/controller/RoomController.java
@@ -66,12 +66,12 @@ public class RoomController {
         try {
             roomService.updateRoom(id, roomDTO);
             MessageDTO response = new MessageDTO();
-            response.setStatus((long) HttpStatus.OK.value());
+            response.setStatus(HttpStatus.OK.value());
             response.setMessage("Mesa actualizada exitosamente");
             return new ResponseEntity<>(response, HttpStatus.OK);
         } catch (RuntimeException e) {
             MessageDTO response = new MessageDTO();
-            response.setStatus((long) HttpStatus.NOT_FOUND.value());
+            response.setStatus(HttpStatus.NOT_FOUND.value());
             response.setMessage(e.getMessage());
             return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
         }
